Remove unnecessary events and refactor.
