+++ components/org.wso2.identity.webhook.common.event.handler/src/main/java/org/wso2/identity/webhook/common/event/handler/api/service/EventProfileManager.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.identity.webhook.common.event.handler.api;
+package org.wso2.identity.webhook.common.event.handler.api.service;
 
 import org.wso2.identity.webhook.common.event.handler.api.model.EventMetadata;
